# docs/swagger.yaml
basePath: /
definitions:
...
  linkpulse_internal_models.ErrorDetails:
    properties:
      code:
        example: INVALID_CREDENTIALS
        type: string
      message:
        example: Invalid email or password
        type: string
    type: object
  linkpulse_internal_models.ErrorResponse:
    properties:
      detail:
        example: The email or password provided is incorrect.
        type: string
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/auth/login
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 401
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Invalid Credentials
        type: string
      type:
        example: https://linkpulse.com/errors/invalid-credentials
        type: string
    type: object
  linkpulse_internal_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  linkpulse_internal_models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  linkpulse_internal_models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      request_id:
        example: uuid-request-id
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  linkpulse_internal_models.UserRegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  linkpulse_internal_models.ValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: The email field is required
        type: string
      rule:
        example: required
        type: string
    type: object
  linkpulse_internal_models.ValidationErrorResponse:
    properties:
      detail:
        example: Validation failed
        type: string
      details:
        items:
          $ref: '#/definitions/linkpulse_internal_models.ValidationError'
        type: array
      error:
        $ref: '#/definitions/linkpulse_internal_models.ErrorDetails'
      instance:
        example: /api/v1/links
        type: string
      request_id:
        example: uuid-request-id
        type: string
      status:
        example: 422
        type: integer
      success:
        example: false
        type: boolean
      title:
        example: Validation Error
        type: string
      type:
        example: https://linkpulse.com/errors/validation-error
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: JyotiBhardwajj
    url: https://github.com/JyotiBhardwajj/LinkPulse
  description: |-
    Production-grade URL shortener with analytics, authentication, and observability.

    ## Authentication
    All protected endpoints require a Bearer JWT in the Authorization header.
    Obtain tokens via POST /api/v1/auth/login.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LinkPulse API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns an access token and refresh token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the provided refresh token, ending the current device session.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - auth
  /api/v1/auth/logout-all:
    post:
      description: Revokes all refresh tokens, ending all device sessions for the
        authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout all sessions
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Accepts a valid refresh token and returns a new access token and
        refresh token pair.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the given email and password.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/linkpulse_internal_models.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ValidationErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/sessions:
    get:
      description: Returns all active device sessions for the currently authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/linkpulse_internal_models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/linkpulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"